requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchaudio
coqui-tts